+++ cmd/uhppote-cli/main.go
@@ -6,10 +6,10 @@ import (
 	"os"
 	"time"
 
-	"github.com/uhppoted/uhppote-cli/commands"
-	"github.com/uhppoted/uhppote-core/types"
-	"github.com/uhppoted/uhppote-core/uhppote"
-	"github.com/uhppoted/uhppoted-lib/config"
+	"codeberg.org/uhppoted/uhppoted-cli/commands"
+	"codeberg.org/uhppoted/uhppoted-core/types"
+	"codeberg.org/uhppoted/uhppoted-core/uhppote"
+	"codeberg.org/uhppoted/uhppoted-lib/config"
 )
 
 var cli = []commands.Command{