+++ commands/get-card.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"os"
 
-	"github.com/uhppoted/uhppote-core/types"
+	"codeberg.org/uhppoted/uhppoted-core/types"
 )
 
 var GetCardCmd = GetCard{}