+++ commands/get-cards.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/uhppoted/uhppote-core/types"
+	"codeberg.org/uhppoted/uhppoted-core/types"
 )
 
 var GetCardsCmd = GetCards{}