+++ commands/get_time_profiles.go
@@ -9,8 +9,8 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/uhppoted/uhppote-core/types"
-	"github.com/uhppoted/uhppoted-lib/config"
+	"codeberg.org/uhppoted/uhppoted-core/types"
+	"codeberg.org/uhppoted/uhppoted-lib/config"
 )
 
 var GetTimeProfilesCmd = GetTimeProfiles{