+++ commands/set_door_control.go
@@ -3,7 +3,7 @@ package commands
 import (
 	"fmt"
 
-	"github.com/uhppoted/uhppote-core/types"
+	"codeberg.org/uhppoted/uhppoted-core/types"
 )
 
 var SetDoorControlCmd = SetDoorControl{}