+++ commands/set_time_profile.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/uhppoted/uhppote-core/types"
-	"github.com/uhppoted/uhppoted-lib/config"
+	"codeberg.org/uhppoted/uhppoted-core/types"
+	"codeberg.org/uhppoted/uhppoted-lib/config"
 )
 
 var SetTimeProfileCmd = SetTimeProfile{}